Modify Charge Category View to show default system charge categories and current provider categories by including selecting NULL hospital_id records from service_category table. Also modified to show edit and delete buttons for non-system default (own provider) charge categories

+++ application/modules/finance/views/service_categoryv2.php
@@ -39,8 +39,12 @@
                                                         <td> <?php echo $category->description; ?></td>
                                                         <?php if ($this->ion_auth->in_group('admin')) { ?>
                                                             <td class="no-print">
-                                                                <a class="btn btn-info btn-xs editbutton" title="<?php echo lang('edit'); ?>" href="finance/editServiceCategory?id=<?php echo $category->id; ?>"><i class="fa fa-edit"></i> </a>
-                                                                <a class="btn btn-danger btn-xs" title="<?php echo lang('delete'); ?>" href="finance/deleteServiceCategory?id=<?php echo $category->id; ?>" onclick="return confirm('Are you sure you want to delete this item?');"><i class="fa fa-trash"></i> </a>
+                                                                <?php if(!empty($category->hospital_id)) { ?>
+                                                                    <a class="btn btn-info btn-xs editbutton" title="<?php echo lang('edit'); ?>" href="finance/editServiceCategory?id=<?php echo $category->id; ?>"><i class="fa fa-edit"></i> </a>
+                                                                <?php } ?>
+                                                                <?php if(!empty($category->hospital_id)) { ?>
+                                                                    <a class="btn btn-danger btn-xs" title="<?php echo lang('delete'); ?>" href="finance/deleteServiceCategory?id=<?php echo $category->id; ?>" onclick="return confirm('Are you sure you want to delete this item?');"><i class="fa fa-trash"></i> </a>
+                                                                <?php } ?>
                                                             </td>
                                                         <?php } ?>
                                                     </tr>
